+++ system/modules/isotope/library/Isotope/CompatibilityHelper.php
